chore: move function to get files from response to utils
