+++ tests/agents/test_quant_agent.py
@@ -2,11 +2,11 @@ import pytest
 
 from langgraph.graph.state import CompiledStateGraph
 
+from src.agents.utils.output_utils import get_all_files_mentioned_in_response
 from src.agents.utils.prompt_utils import extract_graph_response_content
 
 from .helpers import (
     california_monthly_sales_in_db,
-    get_all_files_mentioned_in_response,
     test_temp_dir,
 )
