+++ tests/agents/test_report_graph_steps.py
@@ -12,9 +12,9 @@ import pandas as pd
 
 from langgraph.graph.state import CompiledStateGraph
 
+from src.agents.utils.output_utils import get_all_files_mentioned_in_response
 from src.configuration.kpis import SalesReportRequest
 from .helpers import (
-    get_all_files_mentioned_in_response,
     test_temp_dir,
     png_file_name,
     csv_file_name,
